refactor: Consolidate utility imports and clean up unused code
- Removed the `ui.ts` utility file and migrated its functionality to `index.ts`.
- Updated various components to use the new utility imports.
- Cleaned up the `LiveSandboxCounter` component by removing the compact version.
- Simplified the `TeamMetricsRequest` type definition by removing the export statement.
- Enhanced the `exponentialSmoothing` function documentation for clarity.
