+++ src/app/dashboard/@header/[teamIdOrSlug]/page.tsx
@@ -3,3 +3,4 @@ export default function TeamDashboardHeaderContent() {
   return null
 }
 
+
