+++ src/features/dashboard/sandboxes/live-counter.tsx
@@ -1,9 +1,8 @@
 'use client'
 
+import { cn } from '@/lib/utils'
 import { formatNumber } from '@/lib/utils/formatting'
-import { cn } from '@/lib/utils/ui'
 import { SemiLiveBadge } from '@/ui/live'
-import { Circle } from 'lucide-react'
 
 interface LiveSandboxCounterProps {
   count: number
@@ -26,30 +25,6 @@ export function LiveSandboxCounter({
       <span className="prose-value-small">{formatNumber(count)}</span>
 
       <span className="prose-label text-fg-tertiary">CONCURRENT SANDBOXES</span>
-    </div>
-  )
-}
-
-// compact version for inline use
-interface CompactLiveSandboxCounterProps {
-  count: number
-  className?: string
-}
-
-export function CompactLiveSandboxCounter({
-  count,
-  className,
-}: CompactLiveSandboxCounterProps) {
-  return (
-    <div
-      className={cn(
-        'inline-flex items-center gap-2 rounded-md border bg-bg-1 px-3 py-1.5',
-        className
-      )}
-    >
-      <Circle className="size-2 fill-accent-positive-highlight text-accent-positive-highlight animate-pulse" />
-      <span className="prose-body-highlight">{formatNumber(count)}</span>
-      <span className="prose-label text-fg-tertiary">LIVE</span>
     </div>
   )
 }
